SERIOUSLY?!?!
This is so deprecated, it’s not even funny. Because the MongoDB driver
is an entirely different project with entirely different namespacing
and an entirely different way to do things, this whole thing so sorely
needs to be rewritten.

Doing an if statement by version number doesn’t seem to be a wise
strategy. Either way, it would seem we’re in this for the long haul…